Finalize the survey export to hexapdf
