--- decidim-forms/app/packs/entrypoints/decidim_questionnaire_answers_pdf.js
@@ -1,1 +1,0 @@
-import "entrypoints/decidim_questionnaire_answers_pdf.scss"
