--- decidim-forms/app/packs/entrypoints/decidim_questionnaire_answers_pdf.scss
@@ -1,1 +1,0 @@
-@import "stylesheets/decidim/forms/questionnaire-answers-pdf.scss";
