--- decidim-forms/app/views/layouts/decidim/forms/admin/questionnaires/questionnaire_answers.html.erb
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="<%= I18n.locale %>" class="no-js">
-  <head>
-      <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-      <meta name="viewport" content="width=device-width">
-      <%= stylesheet_pack_tag "decidim_questionnaire_answers_pdf" %>
-      <title><%= @title %></title>
-    </head>
-  <body>
-    <%= yield %>
-  </body>
-</html>
