+++ decidim-forms/config/assets.rb
@@ -5,6 +5,5 @@ base_path = File.expand_path("..", __dir__)
 Decidim::Webpacker.register_path("#{base_path}/app/packs")
 Decidim::Webpacker.register_entrypoints(
   decidim_forms: "#{base_path}/app/packs/entrypoints/decidim_forms.js",
-  decidim_forms_admin: "#{base_path}/app/packs/entrypoints/decidim_forms_admin.js",
-  decidim_questionnaire_answers_pdf: "#{base_path}/app/packs/entrypoints/decidim_questionnaire_answers_pdf.js"
+  decidim_forms_admin: "#{base_path}/app/packs/entrypoints/decidim_forms_admin.js"
 )
